+++ assets/js/components/sidebar/sidebar.html
@@ -14,14 +14,14 @@
 			</a>
 		</li>
 		<li class="nav-item mt-2">
-			<a href="" class="nav-link rounded-pill">
+			<a href="/pages/activity-tracker.html" class="nav-link rounded-pill">
 				<!--BUG: Cannot insert SVG here-->	
 				<img class = "nav-link-icon" src ="../../../../assets/images/icons/chart-line-solid.svg" alt = "An icon that is a line chart"/><span>Activity Tracker</span>								
 				
 			</a>
 		</li>
 		<li class="nav-item mt-2">
-			<a href="" class="nav-link rounded-pill">
+			<a href="/pages/individual-contributions.html" class="nav-link rounded-pill">
 				<!--BUG: Cannot insert SVG here-->	
 				<img class = "nav-link-icon" src ="../../../../assets/images/icons/user-solid.svg" alt = "An icon representing a user"/><span>Individual Contributions</span>
