+++ pages/activity-tracker.html
@@ -1,0 +1,54 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0/dist/css/bootstrap.min.css" rel="stylesheet"
+            integrity="sha384-wEmeIV1mKuiNpC+IOBjI7aAzPcEZeedi5yW5f2yOq55WWLwNGmvvx4Um1vskeMj0" crossorigin="anonymous" />
+        <link rel="stylesheet" href="/assets/css/dashboard.css">
+        	 <!-- WEBSITE TITLE-->
+	 <title>JDG Activity Tracker</title>
+
+     <!-- FAVICON -->
+     <link rel="icon" type="image/jpg" href="/assets/images/favicon.ico" />
+    </head>
+    <body>
+        <page class="d-flex">
+            <sidebar-component></sidebar-component>
+            <main class="w-100">
+                <!-- Contains stuff from js - don't delete -->
+			<!-- <div id='card-container' class='d-flex'>
+            </div> -->
+                <div class="row w-100">
+                    <div class="col-4">
+                        <canvas id="myChart" width="600" height="400"></canvas>
+                    </div>
+                </div>
+                <div style="margin-top: 40px;">
+                    <footer-component></footer-component>
+                </div>
+            </main>
+        </page>
+    
+        <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0/dist/js/bootstrap.bundle.min.js"
+            integrity="sha384-p34f1UUtsS3wqzfto5wAAmdvj+osOnFyQFpp4Ua3gs/ZVWx6oOypYoCJhGGScy+8"
+            crossorigin="anonymous"></script>
+        <script src="https://cdn.jsdelivr.net/npm/chart.js@3.2.1/dist/chart.min.js"></script>
+        <script src="../assets/js/components/component-map.js" type="module" defer></script>
+        <script src="../assets/js/activity-tracker.js" type="module"></script>
+    </body>
+
+</html>
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
