+++ pages/individual-contributions.html
@@ -6,11 +6,10 @@
 	<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0/dist/css/bootstrap.min.css" rel="stylesheet"
 		integrity="sha384-wEmeIV1mKuiNpC+IOBjI7aAzPcEZeedi5yW5f2yOq55WWLwNGmvvx4Um1vskeMj0" crossorigin="anonymous" />
 	<link rel="stylesheet" href="/assets/css/dashboard.css">
-		 <!-- WEBSITE TITLE-->
-		 <title>JDG Dashboard</title>
+    <title>JDG Individual Contributions</title>
 
-		 <!-- FAVICON -->
-		 <link rel="icon" type="image/jpg" href="/assets/images/favicon.ico" />
+    <!-- FAVICON -->
+    <link rel="icon" type="image/jpg" href="/assets/images/favicon.ico" />
 </head>
 
 <body>
@@ -20,8 +19,26 @@
 			<!-- Contains stuff from js - don't delete -->
 			<!-- <div id='card-container' class='d-flex'>
             </div> -->
-			<ul class="nav nav-tabs" id="repo-tabs" role="tablist"></ul>
-			<div class="tab-content" id="repo-tabs-content"></div>
+			<div class="row w-100">
+				<div class="col-4">
+					<table class="table table-striped">
+						<thead>
+							<tr>
+								<th>Name</th>
+								<th>Contributions</th>
+								<th></th>
+							</tr>
+						</thead>
+						<tbody id="contributors-container"></tbody>
+					</table>
+				</div>
+				<div class="col-4">
+					<canvas id="contributor-chart" width="600" height="400"></canvas>
+				</div>
+			</div>
+			<div style="margin-top: 40px;">
+				<footer-component></footer-component>
+			</div>
 		</main>
 	</page>
 
@@ -30,7 +47,7 @@
 		crossorigin="anonymous"></script>
 	<script src="https://cdn.jsdelivr.net/npm/chart.js@3.2.1/dist/chart.min.js"></script>
 	<script src="../assets/js/components/component-map.js" type="module" defer></script>
-	<script src="../assets/js/dashboard.js" type="module"></script>
+	<script src="../assets/js/individual-contributions.js" type="module"></script>
 	<script src="/assets//js/homepage-scripts.js"></script>
 </body>
