+++ pages/repositories.html
@@ -29,7 +29,7 @@
 	 />
  
 	 <!-- WEBSITE TITLE-->
-	 <title>JDG Dashboard</title>
+	 <title>JDG Repositories</title>
 
 	  <!-- FAVICON -->
 	  <link rel="icon" type="image/jpg" href="/assets/images/favicon.ico" />
